refactor: move solvePuzzle function to own module solve_puzzle.js
